+++ addons/es/verifactu/bill_test.go
@@ -7,7 +7,6 @@ import (
 	"github.com/invopop/gobl/addons/es/verifactu"
 	"github.com/invopop/gobl/bill"
 	"github.com/invopop/gobl/cal"
-	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/num"
 	"github.com/invopop/gobl/org"
 	"github.com/invopop/gobl/tax"
@@ -152,9 +151,9 @@ func testInvoiceStandard(t *testing.T) *bill.Invoice {
 				},
 			},
 		},
-		Notes: []*cbc.Note{
+		Notes: []*org.Note{
 			{
-				Key:  cbc.NoteKeyGeneral,
+				Key:  org.NoteKeyGeneral,
 				Text: "This is a test invoice",
 			},
 		},