+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Added
+
+- `untdid`: 4451 - `untdid-text-subject` (Text Subject Qualifer) catalogue.
+
+### Changed
+
+- `org`: Moved `cbc.Note` to `org.Note` in order to include extensions.
+- `tax`: using `ScenarioNote` instead of `cbc.Note` to avoid cyclic dependency issues.
+
 ## [v0.207.0] - 2024-12-12
 
 ### Added